chore: names, icon, error catch
